debugulator: some refactoring
