+++ src/vcf/debugulator.cpp
@@ -22,58 +22,8 @@ namespace ebi
   {
     namespace debugulator
     {
-      namespace po = boost::program_options;
-
-      po::options_description build_command_line_options()
-      {
-          po::options_description description("Usage: vcf-debugulator [OPTIONS] [< input_file]\nAllowed options");
-
-          description.add_options()
-                             ("help,h", "Display this help")
-                             ("input,i", po::value<std::string>()->default_value("stdin"),
-                              "Path to the input VCF file, or stdin")
-                             ("errors,e", po::value<std::string>()->required(),
-                              "Path to the errors DB from the input VCF file")
-                             ("level,l", po::value<std::string>()->default_value("warning"),
-                              "Validation level (error, warning, stop)")
-                             ("output,o", po::value<std::string>()->default_value("stdout"),
-                              "write to a file or stdout");
-
-          return description;
-      }
-
-      int check_command_line_options(po::variables_map const &vm, po::options_description const &desc)
-      {
-          if (vm.count("help")) {
-            std::cout << desc << std::endl;
-            return -1;
-          }
-
-          std::string level = vm["level"].as<std::string>();
-          if (level != "error" && level != "warning" && level != "stop") {
-            std::cout << desc << std::endl;
-            std::cout << "Please choose one of the accepted validation levels" << std::endl;
-            return 1;
-          }
-
-          return 0;
-      }
-
-      ValidationLevel get_validation_level(std::string const &level_str)
-      {
-          if (level_str == "error") {
-            return ValidationLevel::error;
-          } else if (level_str == "warning") {
-            return ValidationLevel::warning;
-          } else if (level_str == "stop") {
-            return ValidationLevel::stop;
-          }
-
-          throw std::invalid_argument{"Please choose one of the accepted validation levels"};
-      }
-
       bool fix_vcf_file(std::istream &input,
-                        ebi::vcf::SqliteReportRW &errorDAO,
+                        ebi::vcf::ReportReader &errorDAO,
                         std::ostream &output)
       {
           std::vector<char> line;
